pyproject.toml
[project]
name = "bookmark-geni-mcp"
version = "1.0.0"
description = "MCP server for processing browser bookmarks, extracting HTML content and metadata, and semantic search"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.1",
    "chromadb>=1.3.4",
    "sentence-transformers==5.1.2",
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.2",
    "html2text>=2025.4.15",
    "pyyaml>=6.0.3",
    "trafilatura>=2.0.0",
    "lxml>=5.3.0",
]

[project.scripts]
bookmark-geni-mcp = "servers.bookmark_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "servers"]
